requirements.txt
elasticsearch
streamlit
python-dateutil
python-dotenv
faker
